fix: CQRS 패턴 적용 (C/Q 분리)
