+++ src/main/java/com/jangburich/presentation/order/controller/OrderCommandController.java
@@ -1,10 +1,10 @@
 package com.jangburich.presentation.order.controller;
 
-import com.jangburich.application.service.order.OrderService;
-import com.jangburich.presentation.order.dto.request.OrderRequest;
-import com.jangburich.presentation.order.dto.response.OrderResponse;
+import com.jangburich.application.service.order.OrderCommandService;
 import com.jangburich.global.payload.Message;
 import com.jangburich.global.payload.ResponseCustom;
+import com.jangburich.presentation.order.dto.request.OrderRequest;
+import com.jangburich.presentation.order.dto.response.OrderResponse;
 import com.jangburich.utils.parser.AuthenticationParser;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -12,13 +12,13 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
-@Tag(name = "Order", description = "Order API")
+@Tag(name = "Order", description = "Order Command Api")
 @RestController
+@RequestMapping("/api/v1/order")
 @RequiredArgsConstructor
-@RequestMapping("/api/v1/orders")
-public class OrderController {
+public class OrderCommandController {
 
-    private final OrderService orderService;
+    private final OrderCommandService orderCommandService;
 
     @Operation(summary = "상품 주문", description = "상품을 주문합니다.")
     @PostMapping
@@ -26,7 +26,7 @@ public class OrderController {
             Authentication authentication,
             @RequestBody OrderRequest orderRequest
     ) {
-        return ResponseCustom.OK(orderService.order(AuthenticationParser.parseUserId(authentication), orderRequest));
+        return ResponseCustom.OK(orderCommandService.order(AuthenticationParser.parseUserId(authentication), orderRequest));
     }
 
     @Operation(summary = "식권 사용", description = "식권을 사용합니다.")
@@ -35,6 +35,6 @@ public class OrderController {
             Authentication authentication,
             @PathVariable Long orderId
     ) {
-        return ResponseCustom.OK(orderService.useMealTicket(AuthenticationParser.parseUserId(authentication), orderId));
+        return ResponseCustom.OK(orderCommandService.useMealTicket(AuthenticationParser.parseUserId(authentication), orderId));
     }
 }
